removed unwanted files and view properties
